test: 컨트롤러 테스트 환경 설정
ControllerTestSupport 작성, @WithMockCustomUser 추가
